got corepattern stuff working 100% for now

+++ lotan/ecfs.c
@@ -627,6 +627,15 @@ static int get_map_count(pid_t pid)
         return lc;
 }
 
+char * get_exe_path(int pid)
+{
+	char *path = xfmtstrdup("/proc/%d/exe", pid);
+	char *ret = (char *)heapAlloc(MAX_PATH);
+	readlink(path, ret, MAX_PATH);
+	free(path);
+	return ret;
+}
+
 /*
  * Get /proc/pid/maps info to create data
  * about stack, heap etc. This can then be
@@ -650,15 +659,12 @@ memdesc_t * build_proc_metadata(pid_t pid, notedesc_t *notedesc)
         
         memset((void *)memdesc->maps, 0, sizeof(mappings_t) * memdesc->mapcount);
         
-	/*
-        if (get_proc_status(notedesc, memdesc) < 0) {
-                printf("[!] failed to get data from /proc/%d/status\n", pid);
-                return NULL;
-        } 
-	*/
 	memdesc->path = exename; // supplied by core_pattern %e
-	printf("exename is: %s\n", memdesc->path);
-        if (get_maps(pid, memdesc->maps, memdesc->path) < 0) {
+       
+	memdesc->exe_path = get_exe_path(pid);
+
+	printf("exe_path: %s\n", memdesc->exe_path);
+	if (get_maps(pid, memdesc->maps, memdesc->path) < 0) {
                 printf("[!] failed to get data from /proc/%d/maps\n", pid);
                 return NULL;
         }
@@ -703,7 +709,8 @@ memdesc_t * build_proc_metadata(pid_t pid, notedesc_t *notedesc)
 static int parse_orig_phdrs(elfdesc_t *elfdesc, memdesc_t *memdesc, notedesc_t *notedesc)
 {
 	int pid = memdesc->task.pid;
-	uint8_t *mem = alloca(8192);
+	int fd;
+	uint8_t *mem;
 	ElfW(Ehdr) *ehdr;
 	ElfW(Phdr) *phdr;
 	ElfW(Addr) text_base = 0;
@@ -728,14 +735,30 @@ static int parse_orig_phdrs(elfdesc_t *elfdesc, memdesc_t *memdesc, notedesc_t *
 		return -1;
 	}
 	
-	if (pid_attach_direct(pid) < 0)
+	/*
+	 * We should avoid using ptrace and it won't work in conjunction with
+	 * the kernels core_pattern piping feature.
+	 */
+	/*
+	if (pid_attach_direct(pid) < 0) {
+		ecfs_print("pid_attach failed\n");
 		return -1;
+	}
 	
 	if (pid_read(pid, (void *)mem, (void *)text_base, 4096) < 0)
 		return -1;
 	
 	if (pid_detach_direct(pid) < 0)
 		return -1;
+	*/
+	/* Instead we use mmap on the original executable file */
+	fd = xopen(memdesc->exe_path, O_RDONLY);
+	mem = mmap(NULL, 8192, PROT_READ, MAP_PRIVATE, fd, 0);
+	if (mem == MAP_FAILED) {
+		perror("mmap");
+		exit(-1);
+	}
+ 
 	/*
 	 * Now get text_base again but from the core file. During a real crashdump
 	 * these values will be the exact same either way.
@@ -792,7 +815,8 @@ static int parse_orig_phdrs(elfdesc_t *elfdesc, memdesc_t *memdesc, notedesc_t *
 				break;
 		}
 	}
-
+	
+	close(fd);
 	return 0;
 }
 
@@ -1851,6 +1875,18 @@ int main(int argc, char **argv)
 				exit(0);
 		}
 	}
+	FILE *fdesc = fopen("/tmp/core.test", "w");	
+	char *pfile = xfmtstrdup("/proc/%d/maps", pid);
+	FILE *fr = fopen(pfile, "r");
+	char tmp[8192];
+	fgets(tmp, sizeof(tmp), fr);
+	fclose(fr);
+	ecfs_print("exename: %s\n", exename);
+	printf("pid: %d\n", pid);
+	fprintf(fdesc, "exename: %s pid: %d\n%s", exename, pid, tmp);
+	fclose(fdesc);
+	
+	opts.logfile = LOGGING_PATH;
 
 	if (opts.use_stdin == 0) {
 		if (corefile == NULL) {
@@ -1884,10 +1920,12 @@ int main(int argc, char **argv)
 		 * open as long as the corefile hasn't been read yet.
 	  	 */
         	if (exename == NULL) {
+			ecfs_print("must specify exename\n");
 			fprintf(stderr, "Must specify exename of process when using stdin mode; supplied by %%e of core_pattern\n");
 			exit(-1);
 		}
 		if (pid == 0) {
+			ecfs_print("must specify a pid\n");
                         printf("Must specify a pid with -p\n");
                         exit(0);
                 }
@@ -1896,11 +1934,13 @@ int main(int argc, char **argv)
                         outfile = xfmtstrdup("%s/ecfs.out", ECFS_CORE_DIR);
                 }
 
+		ecfs_print("calling build_proc_metadata\n");
 		memdesc = build_proc_metadata(pid, notedesc);
         	if (memdesc == NULL) {
                 	fprintf(stderr, "Failed to retrieve process metadata\n");
                 	exit(-1);
         	}
+		ecfs_print("succeeded in calling proc_metadata\n");
 		memdesc->task.pid = pid;
 	}
 
@@ -1935,6 +1975,7 @@ int main(int argc, char **argv)
 	 */
 	notedesc = (notedesc_t *)parse_notes_area(elfdesc);
 	if (notedesc == NULL) {
+		ecfs_print("Failed to parse notes\n");
 		fprintf(stderr, "Failed to parse ELF notes segment\n");
 		exit(-1);
 	}
@@ -1950,7 +1991,6 @@ int main(int argc, char **argv)
 	if (opts.use_stdin == 0) {
 		exename = notedesc->psinfo->pr_fname;
 		pid = pid ? pid : notedesc->prstatus->pr_pid;
-		printf("pid: %d\n", pid);
 		memdesc = build_proc_metadata(pid, notedesc);
         	if (memdesc == NULL) {
                 	fprintf(stderr, "Failed to retrieve process metadata\n");
@@ -1999,6 +2039,7 @@ int main(int argc, char **argv)
 	 * Out of the parsed NT_FILES get a list of which ones are
 	 * shared libraries so we can create shdrs for them.
 	 */
+	ecfs_print("lookup_lib_maps\n");
 	notedesc->lm_files = (struct lib_mappings *)heapAlloc(sizeof(struct lib_mappings));
 	lookup_lib_maps(elfdesc, memdesc, notedesc->nt_files, notedesc->lm_files);
 	
@@ -2011,6 +2052,7 @@ int main(int argc, char **argv)
 	 * data and code is from the dynamic segment by parsing
 	 * it by D_TAG values.
 	 */
+	ecfs_print("extracting dyninfo\n");
 	ret = extract_dyntag_info(handle);
 	if (ret < 0) {
 		fprintf(stderr, "Failed to extract dynamic segment information\n");
@@ -2020,14 +2062,15 @@ int main(int argc, char **argv)
 	 * Convert the core file into an actual ECFS file and write it
 	 * to disk.
 	 */
-	printf("core2ecfs(%s, handle)\n", outfile);
+	ecfs_print("core2ecfs\n");
 	ret = core2ecfs(outfile, handle);
 	if (ret < 0) {
 		fprintf(stderr, "Failed to transform core file '%s' into ecfs\n", argv[2]);
 		exit(-1);
 	}
 	
-	unlink(elfdesc->path);
+	if (opts.use_stdin)
+		unlink(elfdesc->path);
 }
